Create attention mask with correct dtype to match scores
torch.full creates float32 by default, which causes dtype promotion
when added to bfloat16 scores. Explicitly pass dtype=scores.dtype
to create the mask in bfloat16.

Fixes RuntimeError: "expected m1 and m2 to have the same dtype,
but got: float != c10::BFloat16"

+++ attention.py
@@ -133,7 +133,7 @@ class Attention(nn.Module):
 
         # Apply causal mask for prefill (seq_len > 1)
         if seq_len > 1:
-            mask = torch.full((seq_len, kv_seq_len), float("-inf"), device=scores.device)  # (seq, seq_total)
+            mask = torch.full((seq_len, kv_seq_len), float("-inf"), device=scores.device, dtype=scores.dtype)  # (seq, seq_total)
             mask = torch.triu(mask, diagonal=kv_seq_len - seq_len + 1)  # (seq, seq_total)
             scores = scores + mask  # (batch, 8, 4, seq, seq_total)
